Re-use existing Thing entities
If a Thing entity exists in SensorThings API with the same name
and description, it will be re-used instead of creating a new
Thing entity with a POST request.

If multiple entities match with the same name and description, then
the first entity in the collection is re-used. This may not be
side-effect free depending on implementation.