Save observations to day-based CSV files
One file for each day (with dates adjusted to UTC), grouped into
year/month directories for each data file.

`get_observations` method was changed to use data_file as an arg
because save_observations needs to know the data_file filename for
storing the observations, and for updating the parsing stats in
the metadata cache file (WIP).