Merge observations into existing saved observations
If a file with observations already is cached, then merge those
existing observations into the ones that were just downloaded.
Duplicate observation rows with the same timestamp are removed;
the first of the duplicates is kept.

Station metadata cache file is updated to log the oldest/newest
known observation in each parsed file. This is separate to the
oldest/newest timestamp that has been uploaded to STA.

Fixes bug where CSV files created by this method did not have
timestamp values in the rows, only observation results.