Standardize on symbolized keys for ECStation
When reading from a file, the keys were strings; but when creating
the metadata in the Class, they were symbols. This caused issues
creating Entities as sometimes they would refer to nil objects.

This change converts to using symbolized keys. For complex key names,
a colon can be followed by a string with spaces and special characters
to make an advanced Ruby symbol.