Fix variability in STA implementations
When POSTing a new entity, some STA implementations return the
entity in the body and others do not (FROST). In the latter case,
an additional GET request is used to retrieve the entity as the
Location header can sometimes not be a normalized URL (for example,
it could be /Things(3)/Datastreams(8) instead of /Datastreams(8)).

This ensures the correct URL is cached locally for the next STA
requests; they require the normalized URL for posting to sub-
collections.