Link ObservationTypes to Units of Measurement
The Unit of Measurement implies a certain type of content in the
Observation, and this can be used to adjust the value of the
"observationType" property for the SensorThings API Datastream
entity.

Additionally, linking to UOM provides some normalization as
currently UOM is the smallest set in the Ontology. (Although
Observed Properties could potentially be smaller.)