Add Store class for observation data
Currently only Environment Canada Stations store observation data
here, and it is not yet read back for observation upload.

DataStore and MetadataStore were modified to use a v2 subdirectory
to not conflict with directory scanning code in the station classes,
and this subdirectory could be removed later once the DataStore
is fully utilized.