Add Station façade class
WORK IN PROGRESS: this breaks the tool at the moment.

This new class acts as a façade to the provider-specific stations, and
will force each station to provide the same four public methods, at
a minimum:

* download_metadata
* upload_metadata
* download_observations
* upload_observations

By standardizing on these methods, the stations will be easier to
refactor their provider-specific code into private methods or shared
modules.