Add overwrite option for metadata
If the metadata has already been cached locally, then do not
overwrite it when re-running a metadata download. If the overwrite
option is enabled, then it will be replaced.

This is useful as some metadata must be manually corrected and I
do not want to re-correct each time the metadata download is tested.