Match ObservedProperties only on name
Existing Observed Properties on the STA server will only be matched
on their name, instead of name and description. This is because
the encoding of descriptions in the URL as a query string does not
always match the entity on the server, causing an ObservedProperty
entity to be constantly re-created instead of re-used.

Matching only on name is fine, as the results are matched individually
anyway.