Re-record HTTP test interactions for EC
This is tricky as some of the test cases assume a blank slate on
the SensorThings API server side, as the ETL behaviour changes if
Entities already exist on the server. To handle this, some tests
are ran individually while resetting the test STA server between
runs. This ONLY needs to be done when the VCR fixtures are being
reset, and a reset is only necessary when there is a larger version
bump or HTTP interactions change.

In this case, the URLs used with Environment Canada changed to now
be HTTPS.