Fix bug with download.php URL
Some files from Data Garrison would fail to update on the remote
server because the download.php URL that needs to be triggered was
formed incorrectly.

For example, a remote file named "MYC" works fine as it has no
space characters.

A remote file named "Test Launch" has to be converted to
"Test%20Launch" for download.php, and to "Test_Launch" for the
TSV file download URL.

This commit fixes the library and tests to match the correct
behaviour.