Add instructions for historical weather data
Starting with documentation for new module to parse/upload data
and metadata for historical KLRS weather station data. Station is
likely Campbell Scientific, but data is parsed from local CSV files
instead of from online data portal.

This is documentation-only commit with *intended* usage/API and will
be modified as necessary when the module is created.