Parse metadata from CSV headers
As data was collected using Campbell Scientific weather stations,
we can re-use that parsing functionality. However in this case we
can auto-populate some of the metadata as we *know* the location
of all the historical KLRS weather data already.

In the next refactor the parsing functionality for the
CampbellScientific module and the KLRSHistoricalWeather module
should be extracted to a Ruby mixin.