Add StationStore class
This is an interface between the Station class and the DataStore
and MetadataStore classes. It will hide some implementation details
(such as the scoping) from the Station class when it accesses its
data and metadata.

Only KLRSHistoricalWeatherStation has been updated to use it.

The local server upload Bash test script has been updated to use
records stored on a RAM disk for speed. These results will be
compared in the future to performance of a database-backed option.