Rebuild VCR replays for station partial downloads
The VCR replays were being matched based on only the HTTP Method
and the URL, meaning requests with *and without* the Range header
would get the same data back (which is incorrect).

This commit fixes that by checking those tests with the headers as
well. The replay tape for these scenarios has been rebuilt too.