Use Parallel to upsert Observations in batches
Using upsert allows ActiveRecord to handle updates to existing
records. Upserting in batches of 1000 is faster than individual
upserts for each record. Parallel upserts of the batches is also
faster than one large upsert to the database. As the input is
expected to not contain any duplicate records, the parallelization
should be safe.