Do not store unit with observations
Unit is stored with datastream in the station metadata. Observations
point to datastreams, so no need to store unit in the observations
data store.

Existing stores that still contain the unit should work fine, with
the unit being ignored. Existing data stores will automatically
be re-written to not include the unit. Postgres data store already
was not storing the unit, so no changes here.