Raise error when using PostgresStore
The EnvironmentCanada module requires additional metadata for
stations that is handled by the Provider class. This code
requires a directory for storing a copy of the stations list CSV
file and is only compatible when using a file-based store. For now,
an error is raised when trying to use Postgres.