Fix matching observations to datastreams
The datastream names have been cached already by the download of
metadata, and newly downloaded observations need to be matched to
the existing datastreams.

Recent changes to this code had the cached datastreams being checked
for the observation's property name, which fails with some data
sources where the observation property name is long (e.g.
"Temperature_9986750_deg_C").

The older (v0.7.0) code would search an array of cached datastream
names *for every observation*, which works but is slow code. This
commit changes it to only do a search of cached datastream names
*once* and store the match in Hash for faster subsequent lookups.