Fix matching of observation properties
This fixes the upload of observations for DataGarrisonStation, as
it had the same bug as download_observations where the property
names for Observations are not a direct match to Datastream names.
Instead, Datastream names must be searched to see if they are
contained inside the Observation property name. The
ObservationPropertyCache class is used to facilitate this.