Download all metadata in Station class
For EnvironmentCanadaStations, additional metadata from a CSV file
is needed to download the observations and metadata in the SWOB-ML
file. This includes the latitude, longitude, and station type
(automatic or manned).

This commit changes it so that the CSV file is downloaded and
parsed as part of the "download metadata" step, and can be ran
completely separate from the Provider class (which used to be the
way to get the CSV file).

This change is necessary, as the Provider class *requires* the file
based metadata store for keeping a local copy of the CSV file. Now
the CSV file must be downloaded once for each station that runs
"download metadata", but this is a fair trade-off to be able to
easily use the Postgres-backed data store as well. (Note that the
Provider classes must be removed before complete support for the
Postgres store is ready.)