Replace properties access in Station class
Instead of having the station attributes as an instance variable,
it should be kept in the metadata state object as that is where
the properties are kept when stored.

EnvironmentCanadaStation has been updated to only load the metadata
from the store at initialization. The Station superclass now has
metadata saving, and subclasses should be using that.